refactored folder layout.
