+++ Testing/README.md
@@ -1,4 +1,19 @@
-K6 Stress Tests
-===
+K6 HealthGateway Tests
+=
+## Smoke Testing
+
+## Load Testing
+
+  > Load testing is primarily concerned with assessing the systems performance, the purpose of stress testing is to assess the availability and stability of the system under heavy load.
+
+## Stress Testing
+
   > Stress Testing is a type of load testing used to determine the limits of the system. The purpose of this test is to verify the stability and reliability of the system under extreme conditions.
 
+## Spike Testing
+
+  > Spike testing is a type of stress testing that immediately overwhelms the system with an extreme surge of load.
+
+## Soak Testing
+
+  > While load testing is primarily concerned with performance assessment, and stress testing is concerned with system stability under extreme conditions, soak testing is concerned with reliability over a long time. The soak test uncovers performance and reliability issues stemming from a system being under pressure for an extended period.
