+++ Testing/src/k6_load.js
@@ -17,7 +17,7 @@
 import http from 'k6/http';
 import { check, sleep } from 'k6';
 import { Rate } from 'k6/metrics';
-import * as common from '../inc/common.js';
+import * as common from './inc/common.js';
 
 export let errorRate = new Rate('errors');
