+++ Testing/src/k6_soak.js
@@ -24,12 +24,9 @@ export let errorRate = new Rate('errors');
 export let options = {
   stages: [
     { duration: '10s', target: 100 }, // below normal load
-    { duration: '1m', target: 100 },
-    { duration: '10s', target: 1400 }, // spike to 1400 users
-    { duration: '3m', target: 1400 }, // stay there 
-    { duration: '10s', target: 100 }, // scale down
-    { duration: '3m', target: 100 },
-    { duration: '10s', target: 0 }, //
+    { duration: '2m', target: 400 },
+    { duration: '3h56m', target: 400 }, // stay at 400 users for hours 'soaking' the system
+    { duration: '2m', target: 0 }, // drop back down 
   ],
 };
 
@@ -55,6 +52,16 @@ export default function () {
   };
 
   let requests = {
+    'comment': {
+      method: 'GET',
+      url: common.CommentUrl + "/" + user.hdid,
+      params: params
+    },
+    'note': {
+      method: 'GET',
+      url: common.NoteUrl + "/" + user.hdid,
+      params: params
+    },
     'patient': {
       method: 'GET',
       url: common.PatientServiceUrl + "/" + user.hdid,
@@ -65,7 +72,6 @@ export default function () {
       url: common.MedicationServiceUrl + "/" + user.hdid,
       params: params
     },
-
     'labs': {
       method: 'GET',
       url: common.LaboratoryServiceUrl + "?hdid=" + user.hdid,
@@ -75,25 +81,24 @@ export default function () {
 
   let responses = http.batch(requests);
 
+  check(responses['note'], {
+    "Note Response Code is 200": (r) => r.status == 200,
+  }) || errorRate.add(1);
+
+  check(responses['comment'], {
+    "Comment Response Code is 200": (r) => r.status == 200,
+  }) || errorRate.add(1);
+
   check(responses['patient'], {
-    "PatientService Response Code is 200": (r) => r.status == 200,
-    "PatientService Response Code is not 504": (r) => r.status != 504,
-    "PatientService Response Code is not 500": (r) => r.status != 500,
-    "PatientService Response Code is not 403": (r) => r.status != 403,
+    "PatientService Response Code is 200": (r) => r.status == 200
   }) || errorRate.add(1);
 
   check(responses['meds'], {
-    "MedicationService Response Code is 200": (r) => r.status == 200,
-    "MedicationService Response Code is not 504": (r) => r.status != 504,
-    "MedicationService Response Code is not 500": (r) => r.status != 500,
-    "MedicationService Response Code is not 403": (r) => r.status != 403,
+    "MedicationService Response Code is 200": (r) => r.status == 200
   }) || errorRate.add(1);
 
   check(responses['labs'], {
-    "LaboratoryService Response Code is 200": (r) => r.status == 200,
-    "LaboratoryService Response Code is not 504": (r) => r.status != 504,
-    "LaboratoryService Response Code is not 500": (r) => r.status != 500,
-    "LaboratoryService Response Code is not 403": (r) => r.status != 403,
+    "LaboratoryService Response Code is 200": (r) => r.status == 200
   }) || errorRate.add(1);
 
   sleep(1);
