--- Testing/src/load/README.md
@@ -1,4 +1,0 @@
-K6 Load Tests for HealthGateway Services
-===
-  > Load testing is primarily concerned with assessing the systems performance, the purpose of stress testing is to assess the availability and stability of the system under heavy load.
-
