--- Testing/src/soak/README.md
@@ -1,9 +1,0 @@
-K6 Spike Test for HealthGateway Services
-===
-  > While load testing is primarily concerned with performance assessment, and stress testing is concerned with system stability under extreme conditions, soak testing is concerned with reliability over a long time. The soak test uncovers performance and reliability issues stemming from a system being under pressure for an extended period.
-
-
-## How To Run
-
-There is a convenient bash script, run/soak.sh.
-
