--- Testing/src/spike/README.md
@@ -1,9 +1,0 @@
-K6 Spike Test for HealthGateway Services
-===
-  > Spike testing is a type of stress testing that immediately overwhelms the system with an extreme surge of load.
-
-
-## How To Run
-
-There is a convenient bash script, run/spike.sh.
-
