--- Testing/src/stress/README.md
@@ -1,4 +1,0 @@
-K6 Stress Tests for HealthGateway Services
-===
-  > Stress Testing is a type of load testing used to determine the limits of the system. The purpose of this test is to verify the stability and reliability of the system under extreme conditions.
-
